+++ examples/lt_ex_hw_wallet.c
@@ -28,69 +28,11 @@
     - Final product usage
 
     Example shows how content of config objects might be used to set different access rights and how chip behaves during the device's lifecycle.
- */
+*/
 
-//-----------------------------------------------------------
-#define TO_PAIRING_KEY_SH0PUB(x)     ((x) << 0)
-#define TO_PAIRING_KEY_H1(x)     ((x) << 8)
-#define TO_PAIRING_KEY_H2(x)     ((x) << 16)
-#define TO_PAIRING_KEY_H3(x)     ((x) << 24)
-
-#define TO_LT_MCOUNTER_0_3(x)    ((x) << 0)
-#define TO_LT_MCOUNTER_4_7(x)    ((x) << 8)
-#define TO_LT_MCOUNTER_8_11(x)   ((x) << 16)
-#define TO_LT_MCOUNTER_12_15(x)  ((x) << 24)
-
-#define TO_ECC_KEY_SLOT_0_7(x)   ((x) << 0)
-#define TO_ECC_KEY_SLOT_8_15(x)  ((x) << 8)
-#define TO_ECC_KEY_SLOT_16_23(x) ((x) << 16)
-#define TO_ECC_KEY_SLOT_24_31(x) ((x) << 24)
-//-----------------------------------------------------------
-#define SESSION_SH0PUB_HAS_ACCESS    (uint8_t)0x01
-#define SESSION_H1_HAS_ACCESS    (uint8_t)0x02
-#define SESSION_H2_HAS_ACCESS    (uint8_t)0x04
-#define SESSION_H3_HAS_ACCESS    (uint8_t)0x08
-//-----------------------------------------------------------
-
-struct lt_config_obj_desc_t {
-    char desc[60];
-    enum CONFIGURATION_OBJECTS_REGS addr;
-};
-
-/** @brief This structure is used glocally in this example to simplify looping
- *         through all config addresses and printing out them into debug */
-static struct lt_config_obj_desc_t config_description[27] = {
-    {"CONFIGURATION_OBJECTS_CFG_START_UP                   ", CONFIGURATION_OBJECTS_CFG_START_UP_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_SLEEP_MODE                 ", CONFIGURATION_OBJECTS_CFG_SLEEP_MODE_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_SENSORS                    ", CONFIGURATION_OBJECTS_CFG_SENSORS_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_DEBUG                      ", CONFIGURATION_OBJECTS_CFG_DEBUG_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_PAIRING_KEY_WRITE      ", CONFIGURATION_OBJECTS_CFG_UAP_PAIRING_KEY_WRITE_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_PAIRING_KEY_READ       ", CONFIGURATION_OBJECTS_CFG_UAP_PAIRING_KEY_READ_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_PAIRING_KEY_INVALIDATE ", CONFIGURATION_OBJECTS_CFG_UAP_PAIRING_KEY_INVALIDATE_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_R_CONFIG_WRITE_ERASE   ", CONFIGURATION_OBJECTS_CFG_UAP_R_CONFIG_WRITE_ERASE_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_R_CONFIG_READ          ", CONFIGURATION_OBJECTS_CFG_UAP_R_CONFIG_READ_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_I_CONFIG_WRITE         ", CONFIGURATION_OBJECTS_CFG_UAP_I_CONFIG_WRITE_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_I_CONFIG_READ          ", CONFIGURATION_OBJECTS_CFG_UAP_I_CONFIG_READ_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_PING                   ", CONFIGURATION_OBJECTS_CFG_UAP_PING_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_R_MEM_DATA_WRITE       ", CONFIGURATION_OBJECTS_CFG_UAP_R_MEM_DATA_WRITE_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_R_MEM_DATA_READ        ", CONFIGURATION_OBJECTS_CFG_UAP_R_MEM_DATA_READ_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_R_MEM_DATA_ERASE       ", CONFIGURATION_OBJECTS_CFG_UAP_R_MEM_DATA_ERASE_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_RANDOM_VALUE_GET       ", CONFIGURATION_OBJECTS_CFG_UAP_RANDOM_VALUE_GET_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_ECC_KEY_GENERATE       ", CONFIGURATION_OBJECTS_CFG_UAP_ECC_KEY_GENERATE_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_ECC_KEY_STORE          ", CONFIGURATION_OBJECTS_CFG_UAP_ECC_KEY_STORE_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_ECC_KEY_READ           ", CONFIGURATION_OBJECTS_CFG_UAP_ECC_KEY_READ_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_ECC_KEY_ERASE          ", CONFIGURATION_OBJECTS_CFG_UAP_ECC_KEY_ERASE_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_ECDSA_SIGN             ", CONFIGURATION_OBJECTS_CFG_UAP_ECDSA_SIGN_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_EDDSA_SIGN             ", CONFIGURATION_OBJECTS_CFG_UAP_EDDSA_SIGN_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_MCOUNTER_INIT          ", CONFIGURATION_OBJECTS_CFG_UAP_MCOUNTER_INIT_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_MCOUNTER_GET           ", CONFIGURATION_OBJECTS_CFG_UAP_MCOUNTER_GET_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_MCOUNTER_UPDATE        ", CONFIGURATION_OBJECTS_CFG_UAP_MCOUNTER_UPDATE_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_MAC_AND_DESTROY        ", CONFIGURATION_OBJECTS_CFG_UAP_MAC_AND_DESTROY_ADDR},
-    {"CONFIGURATION_OBJECTS_CFG_UAP_SERIAL_CODE_GET        ", CONFIGURATION_OBJECTS_CFG_UAP_SERIAL_CODE_GET_ADDR}
-};
 
 /** @brief  User Conﬁguration Objects - values set here reflect the concept of a generic hardware wallet */
-static uint32_t TROPIC01_config[27] = {
+struct lt_config_t config = {.obj={
     //-------CONFIGURATION_OBJECTS_CFG_START_UP------------------------------------
     // Enable checks on boot
         (CONFIGURATION_OBJECTS_CFG_START_UP_MBIST_DIS_MASK |
@@ -129,17 +71,17 @@ static uint32_t TROPIC01_config[27] = {
     0,
     //------- CONFIGURATION_OBJECTS_CFG_UAP_PAIRING_KEY_READ ----------------------
     // All sessions can read pairing keys
-        (TO_PAIRING_KEY_SH0PUB(SESSION_SH0PUB_HAS_ACCESS | SESSION_H1_HAS_ACCESS | SESSION_H2_HAS_ACCESS | SESSION_H3_HAS_ACCESS) |
-         TO_PAIRING_KEY_H1(SESSION_SH0PUB_HAS_ACCESS | SESSION_H1_HAS_ACCESS | SESSION_H2_HAS_ACCESS | SESSION_H3_HAS_ACCESS) |
-         TO_PAIRING_KEY_H2(SESSION_SH0PUB_HAS_ACCESS | SESSION_H1_HAS_ACCESS | SESSION_H2_HAS_ACCESS | SESSION_H3_HAS_ACCESS) |
-         TO_PAIRING_KEY_H3(SESSION_SH0PUB_HAS_ACCESS | SESSION_H1_HAS_ACCESS | SESSION_H2_HAS_ACCESS | SESSION_H3_HAS_ACCESS)),
+        (TO_PAIRING_KEY_SH0(SESSION_SH0_HAS_ACCESS | SESSION_SH1_HAS_ACCESS | SESSION_SH2_HAS_ACCESS | SESSION_SH3_HAS_ACCESS) |
+         TO_PAIRING_KEY_SH1(SESSION_SH0_HAS_ACCESS | SESSION_SH1_HAS_ACCESS | SESSION_SH2_HAS_ACCESS | SESSION_SH3_HAS_ACCESS) |
+         TO_PAIRING_KEY_SH2(SESSION_SH0_HAS_ACCESS | SESSION_SH1_HAS_ACCESS | SESSION_SH2_HAS_ACCESS | SESSION_SH3_HAS_ACCESS) |
+         TO_PAIRING_KEY_SH3(SESSION_SH0_HAS_ACCESS | SESSION_SH1_HAS_ACCESS | SESSION_SH2_HAS_ACCESS | SESSION_SH3_HAS_ACCESS)),
     //------- CONFIGURATION_OBJECTS_CFG_UAP_PAIRING_KEY_INVALIDATE ----------------
     // Pairing key SH0PUB can be invalidated only from session SH0PUB
     // H3 session can invalidate pairing key H1 H2 and H3
-        (TO_PAIRING_KEY_SH0PUB(SESSION_SH0PUB_HAS_ACCESS ) |
-         TO_PAIRING_KEY_H1(SESSION_H3_HAS_ACCESS ) |
-         TO_PAIRING_KEY_H2(SESSION_H3_HAS_ACCESS ) |
-         TO_PAIRING_KEY_H3(SESSION_H3_HAS_ACCESS)),
+        (TO_PAIRING_KEY_SH0(SESSION_SH0_HAS_ACCESS ) |
+         TO_PAIRING_KEY_SH1(SESSION_SH3_HAS_ACCESS ) |
+         TO_PAIRING_KEY_SH2(SESSION_SH3_HAS_ACCESS ) |
+         TO_PAIRING_KEY_SH3(SESSION_SH3_HAS_ACCESS)),
     //------- CONFIGURATION_OBJECTS_CFG_UAP_R_CONFIG_WRITE_ERASE ------------------
     // Reset value, not used currently
         0x000000ff,
@@ -154,7 +96,7 @@ static uint32_t TROPIC01_config[27] = {
         0x0000ffff,
     //------- CONFIGURATION_OBJECTS_CFG_UAP_PING ----------------------------------
     // Ping command is available for all session keys
-        (SESSION_SH0PUB_HAS_ACCESS | SESSION_H1_HAS_ACCESS | SESSION_H2_HAS_ACCESS | SESSION_H3_HAS_ACCESS),
+        (SESSION_SH0_HAS_ACCESS | SESSION_SH1_HAS_ACCESS | SESSION_SH2_HAS_ACCESS | SESSION_SH3_HAS_ACCESS),
     //------- CONFIGURATION_OBJECTS_CFG_UAP_R_MEM_DATA_WRITE ----------------------
     // Reset value, not used currently
         0xffffffff,
@@ -165,193 +107,71 @@ static uint32_t TROPIC01_config[27] = {
     // Reset value, not used currently
         0xffffffff,
     //------- CONFIGURATION_OBJECTS_CFG_UAP_RANDOM_VALUE_GET ----------------------
-        (SESSION_SH0PUB_HAS_ACCESS | SESSION_H1_HAS_ACCESS | SESSION_H2_HAS_ACCESS | SESSION_H3_HAS_ACCESS),
+        (SESSION_SH0_HAS_ACCESS | SESSION_SH1_HAS_ACCESS | SESSION_SH2_HAS_ACCESS | SESSION_SH3_HAS_ACCESS),
     //------- CONFIGURATION_OBJECTS_CFG_UAP_ECC_KEY_GENERATE ----------------------
     // No session can generate key in ecc key slot 0-7
     // H3 can generate key in slots 8-31
         (TO_ECC_KEY_SLOT_0_7(0)                  |
-         TO_ECC_KEY_SLOT_8_15(SESSION_H3_HAS_ACCESS)  |
-         TO_ECC_KEY_SLOT_16_23(SESSION_H3_HAS_ACCESS) |
-         TO_ECC_KEY_SLOT_24_31(SESSION_H3_HAS_ACCESS)),
+         TO_ECC_KEY_SLOT_8_15(SESSION_SH3_HAS_ACCESS)  |
+         TO_ECC_KEY_SLOT_16_23(SESSION_SH3_HAS_ACCESS) |
+         TO_ECC_KEY_SLOT_24_31(SESSION_SH3_HAS_ACCESS)),
     //------- CONFIGURATION_OBJECTS_CFG_UAP_ECC_KEY_STORE -------------------------
     // H1 can store key into ecc key slot 0-7
     // H3 can store key into ecc key slot 8-31
-        (TO_ECC_KEY_SLOT_0_7(SESSION_H1_HAS_ACCESS)   |
-         TO_ECC_KEY_SLOT_8_15(SESSION_H3_HAS_ACCESS)  |
-         TO_ECC_KEY_SLOT_16_23(SESSION_H3_HAS_ACCESS) |
-         TO_ECC_KEY_SLOT_24_31(SESSION_H3_HAS_ACCESS)),
+        (TO_ECC_KEY_SLOT_0_7(SESSION_SH1_HAS_ACCESS)   |
+         TO_ECC_KEY_SLOT_8_15(SESSION_SH3_HAS_ACCESS)  |
+         TO_ECC_KEY_SLOT_16_23(SESSION_SH3_HAS_ACCESS) |
+         TO_ECC_KEY_SLOT_24_31(SESSION_SH3_HAS_ACCESS)),
     //------- CONFIGURATION_OBJECTS_CFG_UAP_ECC_KEY_READ --------------------------
     // All keys can read pubkey
-        (TO_ECC_KEY_SLOT_0_7(  SESSION_SH0PUB_HAS_ACCESS | SESSION_H1_HAS_ACCESS | SESSION_H2_HAS_ACCESS | SESSION_H3_HAS_ACCESS) |
-         TO_ECC_KEY_SLOT_8_15( SESSION_SH0PUB_HAS_ACCESS | SESSION_H1_HAS_ACCESS | SESSION_H2_HAS_ACCESS | SESSION_H3_HAS_ACCESS) |
-         TO_ECC_KEY_SLOT_16_23(SESSION_SH0PUB_HAS_ACCESS | SESSION_H1_HAS_ACCESS | SESSION_H2_HAS_ACCESS | SESSION_H3_HAS_ACCESS) |
-         TO_ECC_KEY_SLOT_24_31(SESSION_SH0PUB_HAS_ACCESS | SESSION_H1_HAS_ACCESS | SESSION_H2_HAS_ACCESS | SESSION_H3_HAS_ACCESS)),
+        (TO_ECC_KEY_SLOT_0_7(  SESSION_SH0_HAS_ACCESS | SESSION_SH1_HAS_ACCESS | SESSION_SH2_HAS_ACCESS | SESSION_SH3_HAS_ACCESS) |
+         TO_ECC_KEY_SLOT_8_15( SESSION_SH0_HAS_ACCESS | SESSION_SH1_HAS_ACCESS | SESSION_SH2_HAS_ACCESS | SESSION_SH3_HAS_ACCESS) |
+         TO_ECC_KEY_SLOT_16_23(SESSION_SH0_HAS_ACCESS | SESSION_SH1_HAS_ACCESS | SESSION_SH2_HAS_ACCESS | SESSION_SH3_HAS_ACCESS) |
+         TO_ECC_KEY_SLOT_24_31(SESSION_SH0_HAS_ACCESS | SESSION_SH1_HAS_ACCESS | SESSION_SH2_HAS_ACCESS | SESSION_SH3_HAS_ACCESS)),
     //------- CONFIGURATION_OBJECTS_CFG_UAP_ECC_KEY_ERASE -------------------------
     // H1 key can erase keys in slot 0-7
     // H3 key can erase key in slot  8-31
-         (TO_ECC_KEY_SLOT_0_7(SESSION_H1_HAS_ACCESS)  |
-         TO_ECC_KEY_SLOT_8_15(SESSION_H3_HAS_ACCESS)  |
-         TO_ECC_KEY_SLOT_16_23(SESSION_H3_HAS_ACCESS) |
-         TO_ECC_KEY_SLOT_24_31(SESSION_H3_HAS_ACCESS)),
+         (TO_ECC_KEY_SLOT_0_7(SESSION_SH1_HAS_ACCESS)  |
+         TO_ECC_KEY_SLOT_8_15(SESSION_SH3_HAS_ACCESS)  |
+         TO_ECC_KEY_SLOT_16_23(SESSION_SH3_HAS_ACCESS) |
+         TO_ECC_KEY_SLOT_24_31(SESSION_SH3_HAS_ACCESS)),
     //------- CONFIGURATION_OBJECTS_CFG_UAP_ECDSA_SIGN ----------------------------
     // H3 can sign with all ECDSA key slots
-        (TO_ECC_KEY_SLOT_0_7(  SESSION_H3_HAS_ACCESS) |
-         TO_ECC_KEY_SLOT_8_15( SESSION_H3_HAS_ACCESS) |
-         TO_ECC_KEY_SLOT_16_23(SESSION_H3_HAS_ACCESS) |
-         TO_ECC_KEY_SLOT_24_31(SESSION_H3_HAS_ACCESS)),
+        (TO_ECC_KEY_SLOT_0_7(  SESSION_SH3_HAS_ACCESS) |
+         TO_ECC_KEY_SLOT_8_15( SESSION_SH3_HAS_ACCESS) |
+         TO_ECC_KEY_SLOT_16_23(SESSION_SH3_HAS_ACCESS) |
+         TO_ECC_KEY_SLOT_24_31(SESSION_SH3_HAS_ACCESS)),
     //------- CONFIGURATION_OBJECTS_CFG_UAP_EDDSA_SIGN ----------------------------
     // H3 can sign with all ECC key slots
-        (TO_ECC_KEY_SLOT_0_7(  SESSION_H3_HAS_ACCESS) |
-         TO_ECC_KEY_SLOT_8_15( SESSION_H3_HAS_ACCESS) |
-         TO_ECC_KEY_SLOT_16_23(SESSION_H3_HAS_ACCESS) |
-         TO_ECC_KEY_SLOT_24_31(SESSION_H3_HAS_ACCESS)),
+        (TO_ECC_KEY_SLOT_0_7(  SESSION_SH3_HAS_ACCESS) |
+         TO_ECC_KEY_SLOT_8_15( SESSION_SH3_HAS_ACCESS) |
+         TO_ECC_KEY_SLOT_16_23(SESSION_SH3_HAS_ACCESS) |
+         TO_ECC_KEY_SLOT_24_31(SESSION_SH3_HAS_ACCESS)),
     //------- CONFIGURATION_OBJECTS_CFG_UAP_MCOUNTER_INIT -------------------------
     // H3 can init all counters
-        (TO_LT_MCOUNTER_0_3(  SESSION_H3_HAS_ACCESS) |
-         TO_LT_MCOUNTER_4_7( SESSION_H3_HAS_ACCESS)  |
-         TO_LT_MCOUNTER_8_11(SESSION_H3_HAS_ACCESS)  |
-         TO_LT_MCOUNTER_12_15(SESSION_H3_HAS_ACCESS)),
+        (TO_LT_MCOUNTER_0_3(  SESSION_SH3_HAS_ACCESS) |
+         TO_LT_MCOUNTER_4_7( SESSION_SH3_HAS_ACCESS)  |
+         TO_LT_MCOUNTER_8_11(SESSION_SH3_HAS_ACCESS)  |
+         TO_LT_MCOUNTER_12_15(SESSION_SH3_HAS_ACCESS)),
     //------- CONFIGURATION_OBJECTS_CFG_UAP_MCOUNTER_GET --------------------------
     // H3 can get all counters
-        (TO_LT_MCOUNTER_0_3(  SESSION_H3_HAS_ACCESS) |
-         TO_LT_MCOUNTER_4_7( SESSION_H3_HAS_ACCESS)  |
-         TO_LT_MCOUNTER_8_11(SESSION_H3_HAS_ACCESS)  |
-         TO_LT_MCOUNTER_12_15(SESSION_H3_HAS_ACCESS)),
+        (TO_LT_MCOUNTER_0_3(  SESSION_SH3_HAS_ACCESS) |
+         TO_LT_MCOUNTER_4_7( SESSION_SH3_HAS_ACCESS)  |
+         TO_LT_MCOUNTER_8_11(SESSION_SH3_HAS_ACCESS)  |
+         TO_LT_MCOUNTER_12_15(SESSION_SH3_HAS_ACCESS)),
     //------- CONFIGURATION_OBJECTS_CFG_UAP_MCOUNTER_UPDATE -----------------------
     // H3 can update all counters
-        (TO_LT_MCOUNTER_0_3(  SESSION_H3_HAS_ACCESS) |
-         TO_LT_MCOUNTER_4_7( SESSION_H3_HAS_ACCESS)  |
-         TO_LT_MCOUNTER_8_11(SESSION_H3_HAS_ACCESS)  |
-         TO_LT_MCOUNTER_12_15(SESSION_H3_HAS_ACCESS)),
+        (TO_LT_MCOUNTER_0_3(  SESSION_SH3_HAS_ACCESS) |
+         TO_LT_MCOUNTER_4_7( SESSION_SH3_HAS_ACCESS)  |
+         TO_LT_MCOUNTER_8_11(SESSION_SH3_HAS_ACCESS)  |
+         TO_LT_MCOUNTER_12_15(SESSION_SH3_HAS_ACCESS)),
     //------- CONFIGURATION_OBJECTS_TODO -----------------------
     // All slots can use MAC-and-destroy
         (0xffffffff),
     //------- CONFIGURATION_OBJECTS_CFG_UAP_SERIAL_CODE_GET -----------------------
         (0),
+}
 };
-
-/**
- * @brief This function establish a secure channel between host MCU and TROPIC01 chip
- *
- * @param h           Device's handle
- * @param shipriv     Host's private pairing key (SHiPUB)
- * @param shipub      Host's public pairing key  (SHiPUB)
- * @param pkey_index  Pairing key's index
- * @return            LT_OK if success, otherwise returns other error code.
- */
-static lt_ret_t verify_chip_and_start_secure_session(lt_handle_t *h, uint8_t *shipriv, uint8_t *shipub, uint8_t pkey_index)
-{
-    lt_ret_t ret = LT_FAIL;
-
-    // This is not used in this example, but let's read it anyway
-    uint8_t chip_id[LT_L2_GET_INFO_CHIP_ID_SIZE] = {0};
-    ret = lt_get_info_chip_id(h, chip_id, LT_L2_GET_INFO_CHIP_ID_SIZE);
-    if (ret != LT_OK) {
-        return ret;
-    }
-
-    // This is not used in this example, but let's read it anyway
-    uint8_t riscv_fw_ver[LT_L2_GET_INFO_RISCV_FW_SIZE] = {0};
-    ret = lt_get_info_riscv_fw_ver(h, riscv_fw_ver, LT_L2_GET_INFO_RISCV_FW_SIZE);
-    if (ret != LT_OK) {
-        return ret;
-    }
-
-    // This is not used in this example, but let's read it anyway
-    uint8_t spect_fw_ver[LT_L2_GET_INFO_SPECT_FW_SIZE] = {0};
-    ret = lt_get_info_spect_fw_ver(h, spect_fw_ver, LT_L2_GET_INFO_SPECT_FW_SIZE);
-    if (ret != LT_OK) {
-        return ret;
-    }
-
-    uint8_t X509_cert[LT_L2_GET_INFO_REQ_CERT_SIZE] = {0};
-    ret = lt_get_info_cert(h, X509_cert, LT_L2_GET_INFO_REQ_CERT_SIZE);
-    if (ret != LT_OK) {
-        return ret;
-    }
-
-    uint8_t stpub[32] = {0};
-    ret = lt_cert_verify_and_parse(X509_cert, LT_L2_GET_INFO_REQ_CERT_SIZE, stpub);
-    if (ret != LT_OK) {
-        return ret;
-    }
-
-    ret = lt_session_start(h, stpub, pkey_index, shipriv, shipub);
-    if (ret != LT_OK) {
-        return ret;
-    }
-
-    return LT_OK;
-}
-
-/**
- * @brief This function reads config objects and print them out
- *
- * @param h           Device's handle
- * @return            LT_OK if success, otherwise returns other error code.
- */
-/*
-static lt_ret_t read_whole_I_config(lt_handle_t *h)
-{
-    lt_ret_t ret;
-
-    LT_LOG("");
-    for (int i=0; i<27;i++) {
-        uint32_t check = 0;
-        ret = lt_i_config_read(h, config_description[i].addr, &check);
-        if(ret != LT_OK) {
-            return ret;
-        }
-        LT_LOG("\t\t%s  %08" PRIX32, config_description[i].desc, check);
-    }
-    LT_LOG("");
-
-    return LT_OK;
-}
-*/
-/**
- * @brief This function reads config objects and print them out
- *
- * @param h           Device's handle
- * @return            LT_OK if success, otherwise returns other error code.
- */
-static lt_ret_t read_whole_R_config(lt_handle_t *h)
-{
-    lt_ret_t ret;
-
-    LT_LOG("");
-    for (int i=0; i<27;i++) {
-        uint32_t check = 0;
-        ret = lt_r_config_read(h, config_description[i].addr, &check);
-        if(ret != LT_OK) {
-            return ret;
-        }
-        LT_LOG("\t\t%s  %08" PRIX32, config_description[i].desc, check);
-    }
-    LT_LOG("");
-
-    return LT_OK;
-}
-
-/**
- * @brief This function writes config objecs of TROPIC01
- *
- * @param h           Device's handle
- * @return            LT_OK if success, otherwise returns other error code.
- */
-static lt_ret_t write_whole_R_config(lt_handle_t *h)
-{
-    lt_ret_t ret;
-
-    for(int i=0; i<27;i++) {
-        ret = lt_r_config_write(h, config_description[i].addr, TROPIC01_config[i]);
-        if(ret != LT_OK) {
-            return ret;
-        }
-    }
-
-    return LT_OK;
-}
 
 /**
  * @brief Initial session, when chip is powered for the first time during manufacturing.
@@ -369,13 +189,26 @@ static int session_initial(void)
     // Establish secure handshake with default H0 pairing keys
     LT_LOG("%s", "Creating session with initial factory keys H0");
     LT_ASSERT(LT_OK, verify_chip_and_start_secure_session(&h, sh0priv, sh0pub, pkey_index_0));
-    //LT_LOG("%s", "TODO: show I config values");
-    //LT_LOG("%s", "R CONFIG read:");
-    //LT_ASSERT(LT_OK, read_whole_R_config(h));
+
+    LT_LOG("%s", "R CONFIG read:");
+    struct lt_config_t r_config_read;
+    LT_ASSERT(LT_OK, read_whole_R_config(&h, &r_config_read));
+    // Print r config
+    for (int i=0; i<27;i++) {
+        printf("\r\n%s, %08" PRIX32, get_conf_desc(i), r_config_read.obj[i]);
+    }
+    printf("\r\n");
+
     LT_LOG("%s", "Writing whole R config");
-    LT_ASSERT(LT_OK, write_whole_R_config(&h));
+    LT_ASSERT(LT_OK, write_whole_R_config(&h, &config));
+
     LT_LOG("%s", "Reading R CONFIG again");
-    read_whole_R_config(&h);
+    read_whole_R_config(&h, &r_config_read);
+    // Print r config
+    for (int i=0; i<27;i++) {
+        printf("\r\n%s, %08" PRIX32, get_conf_desc(i), r_config_read.obj[i]);
+    }
+    printf("\r\n");
 
     // Writing pairing keys 1-3
     LT_LOG("%s", "Writing pairing key H1");